v1.7.3
Added Interactive CPX model
