+++ Adafruit AR/QRViewController.swift
@@ -13,8 +13,6 @@ import AVFoundation
 
 class QRViewController: UIViewController, ARSCNViewDelegate {
     
-    //test will be in the view did load
-    
     
     // AR Objects
 
@@ -52,6 +50,9 @@ class QRViewController: UIViewController, ARSCNViewDelegate {
     var crickitStopText = SCNNode()
     var crickitSwitch = Bool()
     var crickitVideoSwitch = Bool()
+    
+    
+
     
     
     //For Grand Central M4 AR Model
@@ -193,6 +194,33 @@ class QRViewController: UIViewController, ARSCNViewDelegate {
     var pyPlayerNode = SKVideoNode()
     
     
+    
+    //CPX Interative model
+    //Circuit Playground Interaction Test
+    var neoPixelButton = SCNNode()
+    var neoPixelInfo = SCNNode()
+    var cpxNeoPixelSwitch = false
+    
+    var ATSButton = SCNNode()
+    var ATSInfo = SCNNode()
+    var cpxATSSwitch = false
+    
+    var crocButton = SCNNode()
+    var crocInfo = SCNNode()
+    var cpxCrocSwitch = false
+    
+    var tempButton = SCNNode()
+    var tempInfo = SCNNode()
+    var cpxTempSwitch = false
+    
+    var speakButton = SCNNode()
+    var speakInfo = SCNNode()
+    var cpxSpeakSwitch = false
+    
+    var micButton = SCNNode()
+    var micInfo = SCNNode()
+    var cpxMicSwitch = false
+    
     // URL's for Product Videos
     let crickitVideoURL = URL(string: "https://s3.amazonaws.com/adafruit-ar/video/product_videos/crickit.mp4")
     
@@ -201,6 +229,7 @@ class QRViewController: UIViewController, ARSCNViewDelegate {
     let cpxVidURL = URL(string: "https://s3.amazonaws.com/adafruit-ar/video/product_videos/cpx.mp4")
     
     let pyVidURL = URL(string: "https://s3.amazonaws.com/adafruit-ar/video/product_videos/pyportal.mp4")
+    
     
     // Transitions
     var fadeOut = SCNAction()
@@ -225,6 +254,15 @@ class QRViewController: UIViewController, ARSCNViewDelegate {
         static let height: CGFloat = 180
     }
     
+
+    func easeOutElastic(_ t: Float) -> Float {
+        let p: Float = 0.3
+        let result = pow(2.0, -10.0 * t) * sin((t - p / 4.0) * (2.0 * Float.pi) / p) + 1.0
+        return result
+    }
+    
+    
+    
     // Highlight Nodes in NeoTrellis
     func highlightNode(_ node: SCNNode) {
 
@@ -239,7 +277,6 @@ class QRViewController: UIViewController, ARSCNViewDelegate {
         cubeArray = [cubeNode1, cubeNode2, cubeNode3, cubeNode4, cubeNode5, cubeNode6, cubeNode7, cubeNode8, cubeNode9, cubeNode10,cubeNode11,cubeNode12, cubeNode13, cubeNode14, cubeNode15, cubeNode16, cubeNode17, cubeNode18, cubeNode19, cubeNode20, cubeNode21, cubeNode22, cubeNode23, cubeNode24, cubeNode25, cubeNode26, cubeNode27, cubeNode28, cubeNode29, cubeNode30,cubeNode31,cubeNode32]
         
     }
-    
     
     // Random Color Function for NeoTrellis
     func randomColor(_ node: SCNNode){
@@ -355,6 +392,8 @@ class QRViewController: UIViewController, ARSCNViewDelegate {
         
     }
     
+    
+    
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
         
@@ -446,1216 +485,7 @@ class QRViewController: UIViewController, ARSCNViewDelegate {
         }
     }
     
-    // Video for HalloWing eyeball animation
-    func halloWingEyeVideo() {
-        //This is the video for HalloWing's Eye Ball animation
-
-        // flip video upside down
-        playerNode.yScale = -1
-        playerNode.name = "playerNode"
-        
-        // create SKScene and set player node on it
-        spriteKitScene = SKScene(size: CGSize(width: AspectRatio.width + 300, height: AspectRatio.height + 300))
-        
-        spriteKitScene.scaleMode = .aspectFill
-        
-        playerNode.position = CGPoint(x: spriteKitScene.size.width/2, y: spriteKitScene.size.height/2)
-        
-        playerNode.size = spriteKitScene.size
-        
-        spriteKitScene.addChild(playerNode)
-        
-        
-        videoPlane.geometry?.firstMaterial?.diffuse.contents = spriteKitScene
-        videoPlane.geometry?.firstMaterial?.isDoubleSided = true
-        
-        playerNode.play()
-        
-        NotificationCenter.default.addObserver(forName: .AVPlayerItemDidPlayToEndTime, object: player.currentItem, queue: nil)
-        { notification in
-            
-            self.player.seek(to: kCMTimeZero)
-            
-            self.player.play()
-            
-            print("reset Video")
-            
-        }
-    }
-    
-    //For Promo Video Playback
-    func playCrickitVideo() {
-        
-        crickitPlayer = AVPlayer(url: crickitVideoURL!)
-        crickitPlayerNode = SKVideoNode(avPlayer: crickitPlayer)
-        
-        // flip video upside down
-        crickitPlayerNode.yScale = -1
-        crickitPlayerNode.name = "CrickitplayerNode"
-        
-        // create SKScene and set player node on it
-        spriteKitScene = SKScene(size: CGSize(width: AspectRatio.width + 300, height: AspectRatio.height + 300))
-        
-        spriteKitScene.scaleMode = .aspectFill
-        
-        crickitPlayerNode.position = CGPoint(x: spriteKitScene.size.width/2, y: spriteKitScene.size.height/2)
-        
-        crickitPlayerNode.size = spriteKitScene.size
-        
-        spriteKitScene.addChild(crickitPlayerNode)
-        
-        
-        crickitVideo.geometry?.firstMaterial?.diffuse.contents = spriteKitScene
-        crickitVideo.geometry?.firstMaterial?.isDoubleSided = true
-        
-        crickitPlayerNode.play()
-        
-        NotificationCenter.default.addObserver(forName: .AVPlayerItemDidPlayToEndTime, object: crickitPlayer.currentItem, queue: nil)
-        { notification in
-            
-            self.crickitPlayer.seek(to: kCMTimeZero)
-            self.crickitPlayer.play()
-        }
-    }
-
-    func playHallowingVideo() {
-        
-        halloWingPlayer = AVPlayer(url: hallowingVideoURL!)
-        halloWingPlayerNode = SKVideoNode(avPlayer: halloWingPlayer)
-        
-        // flip video upside down
-        halloWingPlayerNode.yScale = -1
-        halloWingPlayerNode.name = "HallowingplayerNode"
-        
-        // create SKScene and set player node on it
-        spriteKitScene = SKScene(size: CGSize(width: AspectRatio.width + 300, height: AspectRatio.height + 300))
-        
-        spriteKitScene.scaleMode = .aspectFill
-        
-        halloWingPlayerNode.position = CGPoint(x: spriteKitScene.size.width/2, y: spriteKitScene.size.height/2)
-        
-        halloWingPlayerNode.size = spriteKitScene.size
-        
-        spriteKitScene.addChild(halloWingPlayerNode)
-        
-        
-        halloWingVideo.geometry?.firstMaterial?.diffuse.contents = spriteKitScene
-        halloWingVideo.geometry?.firstMaterial?.isDoubleSided = true
-        
-        halloWingPlayerNode.play()
-        
-        NotificationCenter.default.addObserver(forName: .AVPlayerItemDidPlayToEndTime, object: halloWingPlayer.currentItem, queue: nil)
-        { notification in
-            
-            self.halloWingPlayer.seek(to: kCMTimeZero)
-            self.halloWingPlayer.play()
-        }
-    }
-    
-    func playCPXVideo() {
-        
-        cpxPlayer = AVPlayer(url: cpxVidURL!)
-        cpxPlayerNode = SKVideoNode(avPlayer: cpxPlayer)
-        
-        // flip video upside down
-        cpxPlayerNode.yScale = -1
-        cpxPlayerNode.name = "CPXplayerNode"
-        
-        // create SKScene and set player node on it
-        spriteKitScene = SKScene(size: CGSize(width: AspectRatio.width + 300, height: AspectRatio.height + 300))
-        
-        spriteKitScene.scaleMode = .aspectFill
-        
-        cpxPlayerNode.position = CGPoint(x: spriteKitScene.size.width/2, y: spriteKitScene.size.height/2)
-        
-        cpxPlayerNode.size = spriteKitScene.size
-        
-        spriteKitScene.addChild(cpxPlayerNode)
-        
-        
-        cpxVideo.geometry?.firstMaterial?.diffuse.contents = spriteKitScene
-        cpxVideo.geometry?.firstMaterial?.isDoubleSided = true
-        
-        cpxPlayerNode.play()
-        
-        NotificationCenter.default.addObserver(forName: .AVPlayerItemDidPlayToEndTime, object: cpxPlayer.currentItem, queue: nil)
-        { notification in
-            
-            self.cpxPlayer.seek(to: kCMTimeZero)
-            self.cpxPlayer.play()
-        }
-    }
-    
-    
-    //For PyPortal Promo Video Playback
-    func playPyVideo() {
-        
-        pyPlayer = AVPlayer(url: pyVidURL!)
-        pyPlayerNode = SKVideoNode(avPlayer: pyPlayer)
-        
-        // flip video upside down
-        pyPlayerNode.yScale = -1
-        pyPlayerNode.name = "PyplayerNode"
-        
-        // create SKScene and set player node on it
-        spriteKitScene = SKScene(size: CGSize(width: AspectRatio.width + 300, height: AspectRatio.height + 300))
-        
-        spriteKitScene.scaleMode = .aspectFill
-        
-        pyPlayerNode.position = CGPoint(x: spriteKitScene.size.width/2, y: spriteKitScene.size.height/2)
-        
-        pyPlayerNode.size = spriteKitScene.size
-        
-        spriteKitScene.addChild(pyPlayerNode)
-        
-        
-        pyVideo.geometry?.firstMaterial?.diffuse.contents = spriteKitScene
-        pyVideo.geometry?.firstMaterial?.isDoubleSided = true
-        
-        pyPlayerNode.play()
-        
-        NotificationCenter.default.addObserver(forName: .AVPlayerItemDidPlayToEndTime, object: pyPlayer.currentItem, queue: nil)
-        { notification in
-            
-            self.pyPlayer.seek(to: kCMTimeZero)
-            self.pyPlayer.play()
-        }
-    }
-    
-
-    
-    //Gesture Recognizers & Actions
-    func registerTapRecognizer() {
-        let pinchGestureRecognizer = UIPinchGestureRecognizer(target: self, action: #selector(pinch))
-        let tapGestureRecognizer =  UITapGestureRecognizer (target:self ,action : #selector (tapped))
-      //  let swipeLeftGesture = UISwipeGestureRecognizer(target: self, action: #selector(handleSwipe))
-        //let swipeRightGesture = UISwipeGestureRecognizer(target: self, action: #selector(handleSwipe))
-
-        self.qrSceneView.addGestureRecognizer(tapGestureRecognizer)
-        self.qrSceneView.addGestureRecognizer(pinchGestureRecognizer)
-
-    }
-
-
-    
-    
-    
-    
-    
-    
-    
-//    //1. Get The Current Rotation From The Gesture
-//    let rotation = Float(gesture.rotation)
-
-//    //2. If The Gesture State Has Changed Set The Nodes EulerAngles.y
-//    if gesture.state == .changed{
-//    isRotating = true
-//    currentNode.eulerAngles.y = currentAngleY + rotation
-//    }
-
-//    //3. If The Gesture Has Ended Store The Last Angle Of The Cube
-
-//    if(gesture.state == .ended) {
-//    currentAngleY = currentNode.eulerAngles.y
-//    isRotating = false
-//    }
-    
-    
-    
-  
-    
-    
-    
-    
-    /// - Parameter gesture: UIRotationGestureRecognizer
-  
-    
-
-//   @objc func handleSwipe(_ gestureRecognize: UISwipeGestureRecognizer) {
-//        
-//        if (gestureRecognize.direction == .left) {
-//        
-//        }
-//        
-//        else if (gestureRecognize.direction == .right) {
-//        
-//        }
-//    }
-    
-    
-    //Tap Gesture
-    @objc func tapped(sender: UITapGestureRecognizer) {
-        
-        let sceneView = sender.view as! ARSCNView
-        
-        let hitResults = sceneView.hitTest(view.center, types: .existingPlaneUsingExtent)
-        
-        let currentTouchPoint = sender.location(in: self.qrSceneView)
-        
-        if !hitResults.isEmpty == false  {
-            print("Surface Touched")
-            
-        } else {
-            
-            print("DID NOT LOAD NODE.")
-        }
-        
-        if sender.state == .ended  {
-            
-            let location: CGPoint = sender.location(in: sceneView)
-            
-            let hits = self.qrSceneView.hitTest(location, options: nil)
-            
-            if hits.count > 0 {
-                
-                
-                let result = hits[0] as! SCNHitTestResult
-                
-                let node = result.node
-                
-                
-                fadeOut = SCNAction.fadeOut(duration: 1)
-                
-                fadeIn = SCNAction.fadeIn(duration: 0.0001)
-                
-                if (node.name) == "crickitLabel2"{
-                    
-                    let alert = UIAlertController(title: "Leaving Adafruit AR", message: "Would you like to visit the Crickit Learn Guide?", preferredStyle: .alert)
-                    
-                    alert.addAction(UIAlertAction(title: "I Don't Like Nice Things", style: .default, handler: nil))
-                    
-                    alert.addAction(UIAlertAction(title: "Lets Go!", style: .default, handler: { (nil) in
-                        
-                        if let url = URL(string: "https://learn.adafruit.com/adafruit-crickit-creative-robotic-interactive-construction-kit") {
-                            UIApplication.shared.open(url, options: [:])
-                        }
-                        
-                    }))
-                    
-                    present(alert, animated: true, completion: nil)
-                    
-                    
-                    print("Launch Circuit Playground Guide")
-                    
-                }
-                
-                    
-                    
-                    //Grand Central Learn Guide
-                
-                else if (node.name) == "Grand Central Learn Guide"{
-                    
-                    print("Grand Central Learn Guide Button Pressed.")
-                    
-                    let alert = UIAlertController(title: "Leaving Adafruit AR", message: "Would you like to visit the Grand Central Learn Guide?", preferredStyle: .alert)
-                    
-                    alert.addAction(UIAlertAction(title: "I Don't Like Nice Things", style: .default, handler: nil))
-                    
-                    alert.addAction(UIAlertAction(title: "Lets Go!", style: .default, handler: { (nil) in
-                        
-                        if let url = URL(string: "https://learn.adafruit.com/adafruit-grand-central") {
-                            UIApplication.shared.open(url, options: [:])
-                        }
-                        
-                    }))
-                    
-                    present(alert, animated: true, completion: nil)
-                    
-                    
-                    print("Launch Grand Central Guide...")
-                    
-                }
-                    
-               else if (node.name) == "Circuit Learn Guide"{
-                    
-                    let alert = UIAlertController(title: "Leaving Adafruit AR", message: "Would you like to visit the Circuit Playground Express Learn Guide?", preferredStyle: .alert)
-                    
-                    alert.addAction(UIAlertAction(title: "I Don't Like Nice Things", style: .default, handler: nil))
-                    
-                    alert.addAction(UIAlertAction(title: "Lets Go!", style: .default, handler: { (nil) in
-                        
-                        if let url = URL(string: "https://learn.adafruit.com/adafruit-circuit-playground-express") {
-                            UIApplication.shared.open(url, options: [:])
-                        }
-                        
-                    }))
-                    
-                    present(alert, animated: true, completion: nil)
-                    
-                    
-                    print("Launching Circuit Playground Guide...")
-                    
-                }
-                    
-                    
-                else if (node.name) == "Crickit video2"{
-                    
-                    if crickitVideoSwitch == false {
-                        
-                        crickitPlayText.isHidden = true
-                        crickitStopText.isHidden = false
-                        crickitVideo.isHidden = false
-                        
-                        playCrickitVideo()
-                        print("on")
-                        
-                    } else{
-                        
-                        print("off")
-                        
-                        crickitPlayText.isHidden = false
-                        crickitStopText.isHidden = true
-                        crickitVideo.isHidden = true
-                        
-                        self.crickitPlayer.seek(to: kCMTimeZero)
-                        crickitPlayerNode.pause()
-                        crickitVideo.geometry?.firstMaterial?.diffuse.contents = nil
-                        
-                    }
-                    
-                    crickitVideoSwitch = !crickitVideoSwitch
-                    
-                }
-                    
-                else if (node.name) == "HallowLabel"{
-                    
-                    
-                    let alert = UIAlertController(title: "Leaving Adafruit AR", message: "Would you like to visit the HalloWing Learn Guide?", preferredStyle: .alert)
-                    
-                    alert.addAction(UIAlertAction(title: "I Don't Like Nice Things", style: .default, handler: nil))
-                    
-                    alert.addAction(UIAlertAction(title: "Lets Go!", style: .default, handler: { (nil) in
-                        
-                        if let url = URL(string: "https://learn.adafruit.com/adafruit-hallowing?view=all") {
-                            UIApplication.shared.open(url, options: [:])
-                        }
-                        
-                    }))
-                    
-                    present(alert, animated: true, completion: nil)
-                    
-                    
-                    print("Launch HalloWing Learn Guide...")
-                    
-                    
-                }
-                    
-                else if (node.name) == "Hallowing video2" {
-                    
-                    
-                    if hallowingVideoSwitch == false {
-                        print("Test1")
-                        
-                        hallowingPlayText.isHidden = true
-                        hallowingStopText.isHidden = false
-                        hallowingLabelDescription.isHidden = true
-                        halloWingVideo.isHidden = false
-                        playHallowingVideo()
-                        hallowingVideoSwitch = true
-                        
-                    } else if hallowingVideoSwitch == true {
-                        print("Test2")
-                        hallowingPlayText.isHidden = false
-                        hallowingStopText.isHidden = true
-                        hallowingLabelDescription.isHidden = false
-                        self.halloWingPlayer.seek(to: kCMTimeZero)
-                        halloWingPlayerNode.pause()
-                        halloWingVideo.geometry?.firstMaterial?.diffuse.contents = nil
-                        
-                        
-                        hallowingVideoSwitch = false
-                        halloWingVideo.isHidden = true
-                    }
-                }
-                    
-                    
-                    
-                else if (node.name) == "Circuit video button2"{
-                    
-                    
-                    if cpxVidSwitch == false {
-                        
-                        
-                        cpxPlayText.isHidden = true
-                        cpxStopText.isHidden = false
-                        cpxLabelDescription.isHidden = true
-                        cpxVideo.isHidden = false
-                        playCPXVideo()
-                        cpxVidSwitch = true
-                        
-                    } else if cpxVidSwitch == true {
-                        
-                        cpxPlayText.isHidden = false
-                        cpxStopText.isHidden = true
-                        cpxLabelDescription.isHidden = false
-                        self.cpxPlayer.seek(to: kCMTimeZero)
-                        cpxPlayerNode.pause()
-                        cpxVideo.geometry?.firstMaterial?.diffuse.contents = nil
-                        
-                        
-                        cpxVidSwitch = false
-                        cpxVideo.isHidden = true
-                    }
-                }
-                    
-                    //- 2
-                else if (node.name) == "Assembly Plane2"{
-                    
-                    print("Testing")
-                    
-                    
-                    if crickitSwitch == false {
-                        
-                        unassembleText.isHidden = false
-                        
-                        assembleText.isHidden = true
-                        
-                        crickitSolo.isHidden = true
-                        
-                        crickitAnimation.isHidden = false
-                        
-                        crickitSwitch = true
-                        
-                    } else if crickitSwitch == true {
-                        
-                        crickitSwitch = false
-                        
-                        assembleText.isHidden = false
-                        
-                        unassembleText.isHidden = true
-                        
-                        crickitSolo.isHidden = false
-                        
-                        crickitAnimation.isHidden = true
-                    }
-                    
-                }
-                    
-                else if (node.name) == "NeoTrellis Learn Guide "{
-                    
-                    
-                    let alert = UIAlertController(title: "Leaving Adafruit AR", message: "Would you like to visit the NeoTrellis M4 Express Learn Guide?", preferredStyle: .alert)
-                    
-                    alert.addAction(UIAlertAction(title: "I Don't Like Nice Things", style: .default, handler: nil))
-                    
-                    alert.addAction(UIAlertAction(title: "Lets Go!", style: .default, handler: { (nil) in
-                        
-                        if let url = URL(string: "https://learn.adafruit.com/adafruit-neotrellis-m4/overview") {
-                            UIApplication.shared.open(url, options: [:])
-                        }
-                    }))
-                    
-                    present(alert, animated: true, completion: nil)
-                    
-                    print("Launching NeoTrellis Guide...")
-                    
-                } else if (node.name) == "NeoTrellis Video"{
-                    
-                    
-                    if neoSwitch == false {
-                        
-                        neoPlayText.isHidden = true
-                        neoStopText.isHidden = false
-                        neoDescription.isHidden = true
-                        neoVideo.isHidden = false
-                        
-                        neoSwitch = true
-                        
-                    } else if neoSwitch == true {
-                        
-                        neoPlayText.isHidden = false
-                        neoStopText.isHidden = true
-                        neoDescription.isHidden = false
-                        
-                        neoSwitch = false
-                        neoVideo.isHidden = true
-                    }
-                    
-                }
-                    
-                    
-                else if (node.name) == "Cube1"{
-                    
-                    if nodeLit1 {
-                        
-                        cubeNode1.geometry?.firstMaterial?.diffuse.contents = UIColor.lightGray
-                        cubeNode1.dehighlight()
-                        nodeLit1 = true
-                        
-                    } else {
-                        randomColor(cubeNode1)
-                        highlightNode(cubeNode1)
-                        nodeLit1 = false
-                    }
-                    
-                }else if (node.name) == "Cube2"{
-                    
-                    if nodeLit2 {
-                        
-                        cubeNode2.geometry?.firstMaterial?.diffuse.contents = UIColor.lightGray
-                        cubeNode2.dehighlight()
-                        nodeLit2 = true
-                        
-                    } else {
-                        randomColor(cubeNode2)
-                        highlightNode(cubeNode2)
-                        nodeLit2 = false
-                    }
-                }
-                    
-                else if (node.name) == "Cube3"{
-                    
-                    if nodeLit3 {
-                        
-                        cubeNode3.geometry?.firstMaterial?.diffuse.contents = UIColor.lightGray
-                        cubeNode3.dehighlight()
-                        nodeLit3 = true
-                        
-                    } else {
-                        randomColor(cubeNode3)
-                        highlightNode(cubeNode3)
-                        nodeLit3 = false
-                    }
-                    
-                }
-                    
-                else if (node.name) == "Cube4"{
-                    
-                    if nodeLit4 {
-                        
-                        cubeNode4.geometry?.firstMaterial?.diffuse.contents = UIColor.lightGray
-                        cubeNode4.dehighlight()
-                        nodeLit4 = true
-                        
-                    } else {
-                        randomColor(cubeNode4)
-                        highlightNode(cubeNode4)
-                        nodeLit4 = false
-                    }
-                    
-                    
-                    
-                }
-                    
-                else if (node.name) == "Cube5"{
-                    
-                    if nodeLit5 {
-                        
-                        cubeNode5.geometry?.firstMaterial?.diffuse.contents = UIColor.lightGray
-                        cubeNode5.dehighlight()
-                        nodeLit5 = true
-                        
-                        
-                    } else {
-                        randomColor(cubeNode5)
-                        highlightNode(cubeNode5)
-                        nodeLit5 = false
-                    }
-                    
-                    
-                    
-                }
-                else if (node.name) == "Cube6"{
-                    
-                    if nodeLit6 {
-                        
-                        cubeNode6.geometry?.firstMaterial?.diffuse.contents = UIColor.lightGray
-                        cubeNode6.dehighlight()
-                        nodeLit6 = true
-                        
-                    } else {
-                        randomColor(cubeNode6)
-                        highlightNode(cubeNode6)
-                        nodeLit6 = false
-                    }
-                    
-                    
-                    
-                }
-                else if (node.name) == "Cube7"{
-                    
-                    if nodeLit7 {
-                        
-                        cubeNode7.geometry?.firstMaterial?.diffuse.contents = UIColor.lightGray
-                        cubeNode7.dehighlight()
-                        nodeLit7 = true
-                        
-                    } else {
-                        randomColor(cubeNode7)
-                        highlightNode(cubeNode7)
-                        
-                        nodeLit7 = false
-                    }
-                    
-                    
-                    
-                }
-                    
-                    
-                else if (node.name) == "Cube8"{
-                    
-                    if nodeLit8 {
-                        
-                        cubeNode8.geometry?.firstMaterial?.diffuse.contents = UIColor.lightGray
-                        cubeNode8.dehighlight()
-                        nodeLit8 = true
-                        
-                    } else {
-                        randomColor(cubeNode8)
-                        highlightNode(cubeNode8)
-                        nodeLit8 = false
-                        
-                    }
-                    
-                    
-                }
-                    
-                else if (node.name) == "Cube9"{
-                    
-                    if nodeLit9 {
-                        
-                        cubeNode9.geometry?.firstMaterial?.diffuse.contents = UIColor.lightGray
-                        cubeNode9.dehighlight()
-                        nodeLit9 = true
-                        
-                    } else {
-                        randomColor(cubeNode9)
-                        highlightNode(cubeNode9)
-                        nodeLit9 = false
-                        
-                    }
-                    
-                    
-                }
-                else if (node.name) == "Cube10"{
-                    
-                    if nodeLit10 {
-                        
-                        cubeNode10.geometry?.firstMaterial?.diffuse.contents = UIColor.lightGray
-                        cubeNode10.dehighlight()
-                        nodeLit10 = true
-                        
-                        
-                    } else {
-                        randomColor(cubeNode10)
-                        highlightNode(cubeNode10)
-                        nodeLit10 = false
-                    }
-                    
-                    
-                }
-                    
-                else if (node.name) == "Cube11"{
-                    
-                    if nodeLit11 {
-                        
-                        cubeNode11.geometry?.firstMaterial?.diffuse.contents = UIColor.lightGray
-                        cubeNode11.dehighlight()
-                        nodeLit11 = true
-                        
-                        
-                    } else {
-                        randomColor(cubeNode11)
-                        highlightNode(cubeNode11)
-                        nodeLit11 = false
-                    }
-                    
-                    
-                }
-                    
-                else if (node.name) == "Cube12"{
-                    
-                    if nodeLit12 {
-                        
-                        cubeNode12.geometry?.firstMaterial?.diffuse.contents = UIColor.lightGray
-                        cubeNode12.dehighlight()
-                        nodeLit12 = true
-                        
-                        
-                    } else {
-                        randomColor(cubeNode12)
-                        highlightNode(cubeNode12)
-                        nodeLit12 = false
-                    }
-                    
-                    
-                }
-                    
-                else if (node.name) == "Cube13"{
-                    
-                    if nodeLit13 {
-                        
-                        cubeNode13.geometry?.firstMaterial?.diffuse.contents = UIColor.lightGray
-                        cubeNode13.dehighlight()
-                        nodeLit13 = true
-                        
-                        
-                    } else {
-                        randomColor(cubeNode13)
-                        highlightNode(cubeNode13)
-                        nodeLit13 = false
-                    }
-                    
-                    
-                }
-                else if (node.name) == "Cube14"{
-                    
-                    if nodeLit14 {
-                        
-                        cubeNode14.geometry?.firstMaterial?.diffuse.contents = UIColor.lightGray
-                        cubeNode14.dehighlight()
-                        nodeLit14 = true
-                        
-                        
-                    } else {
-                        randomColor(cubeNode14)
-                        highlightNode(cubeNode14)
-                        nodeLit14 = false
-                    }
-                    
-                    
-                }
-                else if (node.name) == "Cube15"{
-                    
-                    if nodeLit15 {
-                        
-                        cubeNode15.geometry?.firstMaterial?.diffuse.contents = UIColor.lightGray
-                        cubeNode15.dehighlight()
-                        nodeLit15 = true
-                        
-                        
-                    } else {
-                        randomColor(cubeNode15)
-                        highlightNode(cubeNode15)
-                        nodeLit15 = false
-                    }
-                    
-                    
-                }
-                else if (node.name) == "Cube16"{
-                    
-                    if nodeLit16 {
-                        
-                        cubeNode16.geometry?.firstMaterial?.diffuse.contents = UIColor.lightGray
-                        cubeNode16.dehighlight()
-                        nodeLit16 = true
-                        
-                        
-                    } else {
-                        randomColor(cubeNode16)
-                        highlightNode(cubeNode16)
-                        nodeLit16 = false
-                    }
-                    
-                    
-                }
-                    
-                else if (node.name) == "Cube17"{
-                    
-                    if nodeLit17 {
-                        
-                        cubeNode17.geometry?.firstMaterial?.diffuse.contents = UIColor.lightGray
-                        cubeNode17.dehighlight()
-                        nodeLit17 = true
-                        
-                        
-                    } else {
-                        randomColor(cubeNode17)
-                        highlightNode(cubeNode17)
-                        nodeLit17 = false
-                    }
-                    
-                    
-                }
-                else if (node.name) == "Cube18"{
-                    
-                    if nodeLit18 {
-                        
-                        cubeNode18.geometry?.firstMaterial?.diffuse.contents = UIColor.lightGray
-                        cubeNode18.dehighlight()
-                        nodeLit18 = true
-                        
-                        
-                    } else {
-                        randomColor(cubeNode18)
-                        highlightNode(cubeNode18)
-                        nodeLit18 = false
-                    }
-                    
-                    
-                }
-                else if (node.name) == "Cube19"{
-                    
-                    if nodeLit19 {
-                        
-                        cubeNode19.geometry?.firstMaterial?.diffuse.contents = UIColor.lightGray
-                        cubeNode19.dehighlight()
-                        nodeLit19 = true
-                        
-                        
-                    } else {
-                        randomColor(cubeNode19)
-                        highlightNode(cubeNode19)
-                        nodeLit19 = false
-                    }
-                    
-                    
-                }
-                    
-                else if (node.name) == "Cube20"{
-                    
-                    if nodeLit20 {
-                        
-                        cubeNode20.geometry?.firstMaterial?.diffuse.contents = UIColor.lightGray
-                        cubeNode20.dehighlight()
-                        nodeLit20 = true
-                        
-                        
-                    } else {
-                        randomColor(cubeNode20)
-                        highlightNode(cubeNode20)
-                        nodeLit20 = false
-                    }
-                    
-                    
-                }
-                    
-                else if (node.name) == "Cube21"{
-                    
-                    if nodeLit21 {
-                        
-                        cubeNode21.geometry?.firstMaterial?.diffuse.contents = UIColor.lightGray
-                        cubeNode21.dehighlight()
-                        nodeLit21 = true
-                        
-                        
-                    } else {
-                        randomColor(cubeNode21)
-                        highlightNode(cubeNode21)
-                        nodeLit21 = false
-                    }
-                    
-                    
-                }
-                else if (node.name) == "Cube22"{
-                    
-                    if nodeLit22 {
-                        
-                        cubeNode22.geometry?.firstMaterial?.diffuse.contents = UIColor.lightGray
-                        cubeNode22.dehighlight()
-                        nodeLit22 = true
-                        
-                        
-                    } else {
-                        randomColor(cubeNode22)
-                        highlightNode(cubeNode22)
-                        nodeLit22 = false
-                    }
-                    
-                    
-                }
-                    
-                else if (node.name) == "Cube23"{
-                    
-                    if nodeLit23 {
-                        
-                        cubeNode23.geometry?.firstMaterial?.diffuse.contents = UIColor.lightGray
-                        cubeNode23.dehighlight()
-                        nodeLit23 = true
-                        
-                        
-                    } else {
-                        randomColor(cubeNode23)
-                        highlightNode(cubeNode23)
-                        nodeLit23 = false
-                    }
-                    
-                    
-                }
-                else if (node.name) == "Cube24"{
-                    
-                    if nodeLit24 {
-                        
-                        cubeNode24.geometry?.firstMaterial?.diffuse.contents = UIColor.lightGray
-                        cubeNode24.dehighlight()
-                        nodeLit24 = true
-                        
-                        
-                    } else {
-                        randomColor(cubeNode24)
-                        highlightNode(cubeNode24)
-                        nodeLit24 = false
-                    }
-                    
-                    
-                }
-                    
-                else if (node.name) == "Cube25"{
-                    
-                    if nodeLit25 {
-                        
-                        cubeNode25.geometry?.firstMaterial?.diffuse.contents = UIColor.lightGray
-                        cubeNode25.dehighlight()
-                        nodeLit25 = true
-                        
-                        
-                    } else {
-                        randomColor(cubeNode25)
-                        highlightNode(cubeNode25)
-                        nodeLit25 = false
-                    }
-                    
-                    
-                }
-                else if (node.name) == "Cube26"{
-                    
-                    if nodeLit26 {
-                        
-                        cubeNode26.geometry?.firstMaterial?.diffuse.contents = UIColor.lightGray
-                        cubeNode26.dehighlight()
-                        nodeLit26 = true
-                        
-                        
-                    } else {
-                        randomColor(cubeNode26)
-                        highlightNode(cubeNode26)
-                        nodeLit26 = false
-                    }
-                    
-                    
-                }
-                    
-                else if (node.name) == "Cube27"{
-                    
-                    if nodeLit27 {
-                        
-                        cubeNode27.geometry?.firstMaterial?.diffuse.contents = UIColor.lightGray
-                        cubeNode27.dehighlight()
-                        nodeLit27 = true
-                        
-                    } else {
-                        randomColor(cubeNode27)
-                        highlightNode(cubeNode27)
-                        nodeLit27 = false
-                    }
-                    
-                    
-                }
-                    
-                else if (node.name) == "Cube28"{
-                    
-                    if nodeLit28 {
-                        
-                        cubeNode28.geometry?.firstMaterial?.diffuse.contents = UIColor.lightGray
-                        cubeNode28.dehighlight()
-                        nodeLit28 = true
-                        
-                    } else {
-                        randomColor(cubeNode28)
-                        highlightNode(cubeNode28)
-                        nodeLit28 = false
-                    }
-                    
-                    
-                }
-                else if (node.name) == "Cube29"{
-                    
-                    if nodeLit29 {
-                        
-                        cubeNode29.geometry?.firstMaterial?.diffuse.contents = UIColor.lightGray
-                        cubeNode29.dehighlight()
-                        nodeLit29 = true
-                        
-                        
-                    } else {
-                        randomColor(cubeNode29)
-                        highlightNode(cubeNode29)
-                        nodeLit29 = false
-                    }
-                    
-                    
-                }
-                    
-                else if (node.name) == "Cube30"{
-                    
-                    if nodeLit30 {
-                        
-                        cubeNode30.geometry?.firstMaterial?.diffuse.contents = UIColor.lightGray
-                        cubeNode30.dehighlight()
-                        nodeLit30 = true
-                        
-                    } else {
-                        randomColor(cubeNode30)
-                        highlightNode(cubeNode30)
-                        nodeLit30 = false
-                    }
-                    
-                    
-                }
-                else if (node.name) == "Cube31"{
-                    
-                    if nodeLit31 {
-                        
-                        cubeNode31.geometry?.firstMaterial?.diffuse.contents = UIColor.lightGray
-                        cubeNode31.dehighlight()
-                        nodeLit31 = true
-                        
-                        
-                    } else {
-                        randomColor(cubeNode31)
-                        highlightNode(cubeNode31)
-                        nodeLit31 = false
-                    }
-                    
-                    
-                }
-                    
-                else if (node.name) == "Cube32"{
-                    
-                    if nodeLit32 {
-                        
-                        cubeNode32.geometry?.firstMaterial?.diffuse.contents = UIColor.lightGray
-                        cubeNode32.dehighlight()
-                        nodeLit32 = true
-                        
-                    } else {
-                        randomColor(cubeNode32)
-                        highlightNode(cubeNode32)
-                        nodeLit32 = false
-                        
-                    }
-                    
-                    
-                }
-                    
-                else if (node.name) == "Pyportal Learn Guide"{
-                    
-                    
-                    let alert = UIAlertController(title: "Leaving Adafruit AR", message: "Would you like to visit the PyPortal Learn Guide?", preferredStyle: .alert)
-                    
-                    alert.addAction(UIAlertAction(title: "I Don't Like Nice Things", style: .default, handler: nil))
-                    
-                    alert.addAction(UIAlertAction(title: "Lets Go!", style: .default, handler: { (nil) in
-                        
-                        if let url = URL(string: "https://learn.adafruit.com/adafruit-pyportal") {
-                            UIApplication.shared.open(url, options: [:])
-                        }
-                    }))
-                    
-                    present(alert, animated: true, completion: nil)
-                    
-                    print("Launching PyPortal Guide...")
-                }
-                    
-                else if (node.name) == "pyVideoLabel" {
-                    
-                    
-                    if pyVidSwitch == false {
-                        print("Test1")
-                        
-                        pyPlayText.isHidden = true
-                        pyStopText.isHidden = false
-                        pyLabelDescription.isHidden = true
-                        pyVideo.isHidden = false
-                        playPyVideo()
-                        pyVidSwitch = true
-                        
-                    } else if pyVidSwitch == true {
-                        print("Test2")
-                        pyPlayText.isHidden = false
-                        pyStopText.isHidden = true
-                        pyLabelDescription.isHidden = false
-                        self.pyPlayer.seek(to: kCMTimeZero)
-                        pyPlayerNode.pause()
-                        pyVideo.geometry?.firstMaterial?.diffuse.contents = nil
-                        
-                        
-                        pyVidSwitch = false
-                        pyVideo.isHidden = true
-                    }
-              
-                
-                }
-                    
-                    
-                else {
-                    print("Nothing Detected.")
-                }
-                
-            }
-        }
-    }
-    
-//
-//    @objc func rotate(sender: UILongPressGestureRecognizer) {
-//        let sceneView = sender.view as! ARSCNView
-//        let holdLocation = sender.location(in: sceneView)
-//        let hitTest = sceneView.hitTest(holdLocation)
-//
-//        if !hitTest.isEmpty {
-//
-//            if sender.state == .began {
-//
-//                if videoWasLoadedToScene == true {
-//
-//                    exitVideoNode()
-//
-//                    fadeInAllLetters()
-//
-//                } else if videoWasLoadedToScene == false {
-//
-//                    let rotation = SCNAction.rotateBy(x: 0, y: CGFloat(360.degreesToRadians), z: 0, duration: 2)
-//
-//                    let forever = SCNAction.repeatForever(rotation)
-//
-//                    adabotNode.runAction(forever)
-//
-//                }
-//
-//            } else if sender.state == .ended {
-//
-//                adabotNode.removeAllActions()
-//            }
-//        }
-//    }
-    
-    @objc func handlePinch(_ pinchRecognize: UIPinchGestureRecognizer) {
-        
-        let scale = Float(pinchRecognize.scale / 10)
-        let currentScale = pyNode.scale.x
-        let newScale = scale / currentScale
-        self.pyNode.scale.x = newScale
-        self.pyNode.scale.y = newScale
-        self.pyNode.scale.z = newScale
-    }
-    
-    
-    
-  
-    //Pinch Gesture
-    @objc func pinch(sender: UIPinchGestureRecognizer) {
-        
-        let sceneView = sender.view as! ARSCNView
-        let pinchLocation = sender.location(in: qrSceneView)
-        let hitResults = sceneView.hitTest(pinchLocation)
-        
-        if !hitResults.isEmpty {
-            
-            let pinchAction = SCNAction.scale(by: sender.scale, duration: 0)
-            print(sender.scale)
-
-            
-            grandCentralNode.runAction(pinchAction)
-            crickitNode.runAction(pinchAction)
-            halloWingNode.runAction(pinchAction)
-            cpxNode.runAction(pinchAction)
-            pyNode.runAction(pinchAction)
-            sender.scale = 1.0
-            
-        }
-    }
-    
-    
+          
     // MARK: - ARSCNViewDelegate
     func renderer(_ renderer: SCNSceneRenderer, nodeFor anchor: ARAnchor) -> SCNNode? {
         //This node is used to place the Plane used to plant the AR models
@@ -1736,7 +566,7 @@ class QRViewController: UIViewController, ARSCNViewDelegate {
                 
                 let scale: Float = 0.05
                 
-                let labelScale: Float = 1.6
+                let labelScale: Float = 1.8
                 
                 let plane = SCNPlane(width: imageAnchor.referenceImage.physicalSize.width, height: imageAnchor.referenceImage.physicalSize.height)
                 
@@ -1764,6 +594,7 @@ class QRViewController: UIViewController, ARSCNViewDelegate {
                 
                 cpxStopText = cpxLabel.childNode(withName: "CStop video", recursively: true)!
                 
+                
                 cpxVideo.isHidden = true
                 
                 cpxStopText.isHidden = true
@@ -1772,9 +603,18 @@ class QRViewController: UIViewController, ARSCNViewDelegate {
                 
                 let CPXScene = SCNScene(named: "art.scnassets/Circuit_Playground.dae")!
                 
+                
                 for child in CPXScene.rootNode.childNodes {
                     cpxNode.addChildNode(child)
                 }
+               
+                
+//                //For Interactive Test
+//                neoPixelButton = cpxNode.childNode(withName: "neopixel Button", recursively: true)!
+//                
+//                neoPixelInfo = cpxNode.childNode(withName: "neopixel Info", recursively: true)!
+//                
+//             //   neoPixelInfo.isHidden = true
                
                 cpxNode.position = SCNVector3(planeNode.position.x, planeNode.position.y, planeNode.position.z)
                 
@@ -1782,7 +622,61 @@ class QRViewController: UIViewController, ARSCNViewDelegate {
                 
                 cpxNode.scale = SCNVector3(x: scale, y: scale, z: scale)
 
-
+               
+                
+                //For Interactive Test
+                
+                //NeoPixels
+                neoPixelButton = cpxNode.childNode(withName: "neopixel_Button", recursively: true)!
+                
+                neoPixelInfo = cpxNode.childNode(withName: "neopixel_Info", recursively: true)!
+                
+                neoPixelInfo.isHidden = true
+               
+                //ATSAMD21
+                ATSButton = cpxNode.childNode(withName: "ATSAMD21_Micro_button", recursively: true)!
+                
+                ATSInfo = cpxNode.childNode(withName: "ATSAMD_info", recursively: true)!
+                
+                
+                ATSInfo.isHidden = true
+                
+                
+                //Croc/Alligator Clip Pads
+                crocButton = cpxNode.childNode(withName: "Croc_Button", recursively: true)!
+                
+                crocInfo = cpxNode.childNode(withName: "Croc_info", recursively: true)!
+                
+                
+                crocInfo.isHidden = true
+                
+               
+                //Temperature Sensor
+                tempButton = cpxNode.childNode(withName: "temp_button", recursively: true)!
+                
+                tempInfo = cpxNode.childNode(withName: "temp_info", recursively: true)!
+                
+                
+                tempInfo.isHidden = true
+                
+                //Speaker
+                speakButton = cpxNode.childNode(withName: "speaker_button", recursively: true)!
+                
+                speakInfo = cpxNode.childNode(withName: "speaker_info", recursively: true)!
+                
+                speakInfo.isHidden = true
+                
+                //Mic
+                micButton = cpxNode.childNode(withName: "mic_button", recursively: true)!
+                
+                micInfo = cpxNode.childNode(withName: "mic_info", recursively: true)!
+                
+                micInfo.isHidden = true
+                
+                
+//                let billboardConstraint = SCNBillboardConstraint()
+//                billboardConstraint.freeAxes = SCNBillboardAxis.Y
+//                micInfo.constraints = [billboardConstraint]
                 
                 planeNode.addChildNode(cpxNode)
                 
@@ -2114,6 +1008,8 @@ class QRViewController: UIViewController, ARSCNViewDelegate {
         
     }
 
+    
+    //FOR UNIT TESTING VVV
     var productsLoaded = Int()
     var productsOutOfStock = Int()
     
@@ -2123,8 +1019,6 @@ class QRViewController: UIViewController, ARSCNViewDelegate {
          productsOutOfStock = 0
         
          aFuctionForReaction()
-    
-    
     }
