+++ Adafruit AR/ViewController+Gesture Recognizers.swift
@@ -14,16 +14,17 @@ extension ViewController {
   
   //Mark:- Gesture Recognizers --------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
   
-  func registerTapRecognizer() {
-    let tapGestureRecognizer =  UITapGestureRecognizer (target:self ,action : #selector (tapped))
-    let pinchGestureRecognizer = UIPinchGestureRecognizer(target: self, action: #selector(pinch))
-    let longPressGestureRecognizer = UILongPressGestureRecognizer(target: self, action: #selector(rotate))
+
+    func registerTapRecognizer() {
+        let tapGestureRecognizer =  UITapGestureRecognizer (target:self ,action : #selector (tapped))
+        let pinchGestureRecognizer = UIPinchGestureRecognizer(target: self, action: #selector(pinch))
+        let longPressGestureRecognizer = UILongPressGestureRecognizer(target: self, action: #selector(rotate))
+        
+        self.sceneView.addGestureRecognizer(tapGestureRecognizer)
+        self.sceneView.addGestureRecognizer(pinchGestureRecognizer)
+        self.sceneView.addGestureRecognizer(longPressGestureRecognizer)
+    }
     
-    self.sceneView.addGestureRecognizer(tapGestureRecognizer)
-    self.sceneView.addGestureRecognizer(pinchGestureRecognizer)
-    self.sceneView.addGestureRecognizer(longPressGestureRecognizer)
-  }
-  
   
   //Gesture Actions
   @objc func tapped(sender: UITapGestureRecognizer) {
